fix: partition PubSubRates for rates type
